Extract out a shared comment
It seems silly to keep repeating the same warning that all of these
examples are not ready for production use. Having to add a similar or
identical comment for each new sample seems excessive. As such, this
commit extracts out the warnings against using these examples verbatim,
making the point that these clients are really just proof-of-concepts.