Update with second-order tokens
This commit includes information about locally-stored second-order REST API authentication tokens. These longer-lifespan tokens will allow readers to return to a previously visited edition directly without having to login for the rest of that day. This is thanks to an improvement in the endpoint authenticating the tokens, which is now able to send a new token directly to the edition for future use.