Provide a use-case
This commit updates the instructions regarding usage of the tokens with a simple use-case. It demonstrates how a server capable of making REST API requests could be trivially augmented to use the tokens. The suggestion is to respond to reader requests in the form of a redirect to the edition URL with the token data inserted into the URL.